+++ api/src/main/java/com/intuit/tank/vm/agent/messages/WatsAgentStartRequest.java
@@ -59,7 +59,7 @@ public class WatsAgentStartRequest implements Serializable {
     private int startUsers;
 
     @XmlElement(name = "userIntervalIncrement", namespace = Namespace.NAMESPACE_V1, required = true, nillable = false)
-    private double userIntervalIncrement;
+    private int userIntervalIncrement;
 
     @XmlElement(name = "agentInstanceNum", namespace = Namespace.NAMESPACE_V1, required = true, nillable = false)
     private int agentInstanceNum;
@@ -154,7 +154,7 @@ public class WatsAgentStartRequest implements Serializable {
         this.userIntervalIncrement = userIntervalIncrement;
     }
 
-    public double getUserIntervalIncrement() {
+    public int getUserIntervalIncrement() {
         return userIntervalIncrement > 0 ? userIntervalIncrement : 1;
     }
