+++ api/src/test/java/com/intuit/tank/vm/agent/messages/AgentTestStartDataCpTest.java
@@ -349,7 +349,7 @@ public class AgentTestStartDataCpTest {
         fixture.setStartUsers(1);
         fixture.setSimulationTime(1L);
 
-        double result = fixture.getUserIntervalIncrement();
+        int result = fixture.getUserIntervalIncrement();
 
         assertEquals(1, result);
     }
@@ -373,7 +373,7 @@ public class AgentTestStartDataCpTest {
         fixture.setStartUsers(1);
         fixture.setSimulationTime(1L);
 
-        double result = fixture.getUserIntervalIncrement();
+        int result = fixture.getUserIntervalIncrement();
 
         assertEquals(1, result);
     }
